server/requirements.txt
asgiref==3.10.0
brotli==1.2.0
certifi==2025.10.5
charset-normalizer==3.4.4
cloudinary==1.44.1
Django==5.2.8
django-admin-shell==2.0.1
django-ckeditor-5==0.2.18
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-js-asset==3.1.2
django-json-widget==2.1.0
djangorestframework==3.16.1
gunicorn==23.0.0
idna==3.11
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
whitenoise==6.11.0
